#4 Adding JS and Python Images
Created classes to run arbitrary methods with arbitrary parameters
passed in. All parameters for python and javascript are named parameters
to reduce invalid calls or mixing the order of the parameters.
The --parameters variable is required to be a quoted JSON string.
There are debian-slim, alpine and micropython images for python
functions.
There are debian-slim and alpine images of Node.js and the alpine image
of Deno for javascript functions.